use new ds config in migrations.
